\name{DateTime-class}
\docType{class}
\alias{DateTime-class}

\title{Class Union "DateTime" }
\description{A class union for the standard data/time classes, \code{"POSIXct"},\code{"POSIXlt"}, and \code{"POSIXt"}.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\seealso{

	Class \code{\linkS4class{GPSTrack}} for a class having a slot of this class.
}
\keyword{classes}
