% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{discovery_query}
\alias{discovery_query}
\title{Query Discovery API}
\usage{
discovery_query(
  offset = 0,
  limit = 100,
  ids = NULL,
  domains = NULL,
  names = NULL,
  categories = NULL,
  tags = NULL,
  types = NULL,
  attribution = NULL,
  license = NULL,
  query = NULL,
  min_should_match = NULL,
  parent_ids = NULL,
  derived_from = NULL,
  provenance = NULL,
  owner = NULL,
  column_names = NULL,
  visibility = NULL,
  audience = NULL,
  published = NULL,
  approval_status = NULL,
  explicitly_hidden = NULL,
  data_json_hidden = NULL,
  derived = NULL,
  order = NULL
)
}
\arguments{
\item{offset}{The initial starting point for paging.}

\item{limit}{Either the maximum number of results to return (up to 10000) or
`NULL`. If an integer, the output will return at most the number of results
in `limit`, started from the `offset`. If `NULL`, the output will return all
results after the `offset`, without any `limit`.}

\item{ids}{The unique four-by-four identifier of an asset. Draft assets can
be identified by appending the four-by-four identifier with either `:draft`
for stories or an integer of the draft number, e.g. `:1`, for non-story
drafts. This parameter supports repeated usage.}

\item{domains}{A comma separated string of canonical domain names. This
parameter does support repeated usage.}

\item{names}{The case-insensitive name of an asset. The output will only
return exact matches. The `query` argument should be used for partial
matches. This parameter supports repeated usage.}

\item{tags}{A single tag. This parameter supports repeated usage.}

\item{types}{A comma separated string of asset types. Any of api, calendar,
chart, datalens, dataset, federated_href, file, filter, form, href, link,
map, measure, story, visualization. This parameter does not support
repeated usage.}

\item{attribution}{The case-sensitive name of the attributing entity.}

\item{license}{The case-sensitive license name.}

\item{query}{A string for full text search.}

\item{min_should_match}{The number or percent of words that must match in `query`.}

\item{parent_ids}{The unique identifier of an asset.}

\item{derived_from}{The unique identifier of an asset.}

\item{provenance}{The provenance, as either 'official' or 'community'.}

\item{owner}{The unique identifier of a user or a team.}

\item{column_names}{The name of a column.}

\item{visibility}{The visibility status, as either 'open' or 'internal'.}

\item{audience}{The audience, as either 'private', 'site' or 'public'.}

\item{published}{A "true" or "false" value, for published or unpublished assets.}

\item{approval_status}{An approval status as enumerated above.}

\item{explicitly_hidden}{A "true" or "false" value, for hidden or unhidden assets.}

\item{data_json_hidden}{A "true" or "false" value, for hidden or unhidden assets.}

\item{derived}{A "true" or "false" value, for derived or base assets.}

\item{order}{The sort order of the results. One of name, owner, dataset_id,
datatype, domain_category, createdAt, updatedAt, page_views_total,
page_views_last_month, page_views_last_week.




#TODO: Finish docs}

\item{catgories}{A single category. This parameter supports repeated usage.}
}
\description{
See https://socratadiscovery.docs.apiary.io/
}
