\name{evalText}
\alias{evalText}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Turn a text string into an evaluated expression }
\description{
 Turns a text string into an evaluated expression.  See the "Computing with Text" chapter for why.
}
\usage{
evalText(text, where = .GlobalEnv)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{text}{ The string. }
  \item{where}{ The environment to use for the computation. }
}

\value{
  Whatever the expression evaluates to.
}

\examples{
##  See the chapter.

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{programming}

