\name{SoDA-Examples}
\alias{SoDA-Examples}
\alias{Examples}
\alias{mars}
\docType{package}
\title{
Example Files for the book "Software for Data Analysis"
}
\description{The SoDA package has a set of files related to examples
  in the book.  These are all found in the directory \code{"Examples"}
  under the installed version of the package.  The files are intended
  to be edited and used as source to experiment with the examples in
  the book.
}
\details{  

  Note that not all the examples shown in the book
  correspond to an explicit example file.  Also, some of the data
  sets used are not explicitly open-source and so are not included in
  this package.  For example, the \code{mars} data exists on the web
  as an Excel file; you need to export that as a \code{".csv"} file to
  run the corresponding examples.

To find all the examples, or examples appearing on a page in the book,
use \code{\link{exampleFiles}}.

See \code{\link{runExample}} to run an example file.


}
\author{
John M. Chambers <jmc@r-project.org>
}
\keyword{programming}
