\name{mergeSections}
\alias{mergeSections}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ Merge documentation sections }
\description{
 Merge documentation sections.
}
\usage{
mergeSections(input, keepFirst = 1, keepLast = 1, sep = "")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{input, keepFirst, keepLast, sep}{ See description in chapter. }
}
\details{
  Support function for \code{\link{promptAll}}
}
\value{
  Merged text.   See chapter.
}
\keyword{internal}

