\name{plot-methods}
\docType{methods}
\alias{plot-methods}
\alias{plot,track,missing-method}
\alias{plot,track3,missing-method}
\alias{show,track-method}
\alias{show,binaryRep-method}
\title{ Methods for Plot and show in the SoDA pacakge }
\description{
 Methods for plotting objects from the track classes.
}
\section{Methods}{
For plot:
\describe{

\item{x = "track", y = "missing"}{ Plots the points on the track.}

\item{x = "track3", y = "missing"}{ Codes slot \code{z} in the plotting character of the inherited method from class \code{track}.}
}
Also  show methods for class \code{track} and class \code{binaryRep}
}
\keyword{methods}
