\name{recoverHandler}
\alias{recoverHandler}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calling handler for recover}
\description{
  A handler to use with function \code{\link{withCallingHandlers}}.
}
\usage{
recoverHandler(condition)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{condition}{ the condition passed in }
}
\value{
  None of interest
}

\keyword{programming}
