\name{SODA_Data}
\alias{SODA_Data}
\docType{data}
\title{
SODA_Data is a data set using in the practical example for SODA_Kriging function
}
\description{
This dataset contains a subset of the Human Breast Cancer data we used in our recent publications downloaded from 10x Genomics online resource.
}
\usage{data("SODA_Data")}
\format{
  A data frame with 500 samples and 1000 genes plus 2 more columns representing Longitude and Latitude of the cells.

}

\source{
Human Breast Cancer: Whole Transcriptome Analysis \cr
Available at: https://www.10xgenomics.com/resources/datasets/human-breast-cancer-whole-transcriptome-analysis-1-standard-1-2-0
}
\references{
Hajihosseini, M., Amini, P., Voicu, D., Dinu, I., & Pyne, S. (2022). Geostatistical Modeling and Heterogeneity Analysis of Tumor Molecular Landscape. Cancers, 14(21), 5235. \cr
Amini, P., Hajihosseini, M., Pyne, S., & Dinu, I. (2022). Geographically Weighted Linear Combination Test for Gene Set Analysis of a Continuous Spatial Phenotype as applied to Intratumor Heterogeneity. bioRxiv.
}
\examples{
data(SODA_Data)

}
\keyword{datasets}
