\name{sotu}
\alias{sotu}
\docType{data}
\title{ Some text variables on the State of the Union Addresses}
\description{This dataset provides some text variables derived from analysis of the text of the State of the Union Addresses.
}
\usage{data(sotu)}
\format{
  A data frame with 217 observations on  15 variables.
  \describe{
    \item{\code{date}}{Date of the speech}
    \item{totallength, wordlength, sentencelength}{ Measures of the lengths in the speeches.}
      \item{america, confidence, democracy, economy, evil, freedom, good, peace, terror, uncertain, war}{Counts of various words in the speeches}
  }
}
\source{  See \url{www.swivel.com}.
}

\keyword{datasets}
