\name{track-class}
\docType{class}
\alias{track-class}
\alias{track3-class}

\title{Class "track" ~~~ }
\description{Planar or 3-dimensional track objects Class \code{track3} contains \code{track}.}
\section{Objects from the Class}{
Objects conceived as being on a plane or in 3-space, representing samples from an observed track.
}
\section{Slots}{
	 \describe{
    \item{\code{x, y, z}:}{Objects of class \code{"numeric"} for the coordinates of the track. }
    
  }
}

\keyword{classes}
