\name{SoDA-Examples}
\alias{SoDA-Examples}
\alias{Examples}
\docType{package}
\title{
Example Files for the book "Software for Data Analysis"
}
\description{The SoDA package has a set of files related to examples in the book.  These are all found in the directory \code{"Examples"} under the installed version of the package.  The files are intended to be edited and used as source to experiment with the examples in the book.
}
\details{The files are not demos, since most of them are not self-contained.  Instead, the idea is that you can extract pieces of code  to run or edit.
To find all the examples, 

\code{list.files(system.file("Examples", package = "SoDA"))}


}
\author{
John M. Chambers <jmc@r-project.org>
}
\keyword{programming}
