\name{SoDA-package}
\alias{SoDA-package}
\alias{SoDA}
\docType{package}
\title{
Functions and Other Code for the book "Software for Data Analysis"
}
\description{
A collection of R functions, some Fortran and C code, a little
miscellaneous code in other languages, and some data sets.  These
arise in various contexts in the book "Software for Data Analysis".
}
\author{
John M. Chambers <jmc@r-project.org>
}
\keyword{ package }
